homework-lesson-5/requirements.txt
langchain>=1.2.0
langchain-core>=1.2.0
langchain-openai>=0.4
langchain-community>=0.4
langchain-classic>=1.0
faiss-cpu
rank_bm25
sentence-transformers
ddgs>=7.0
trafilatura>=2.0.0
pypdf
pydantic>=2.12.0
pydantic-settings>=2.12.0
